+++ src/index.ts
@@ -2,8 +2,5 @@ import './styles/global.scss';
 
 export * from './components/Popup';
 export * from './components/Surface';
-export * from './surveys/CSAT2Survey';
-export * from './surveys/CSAT5Survey';
-export * from './surveys/CES7Survey';
-export * from './surveys/NPS10Survey';
+export * from './surveys/Survey';
 export * from './types';