+++ src/surveys/CES7Survey/index.ts
@@ -1,1 +1,0 @@
-export { CES7Survey, type CES7SurveyProps } from './CES7Survey';