+++ src/surveys/CSAT2Survey/index.ts
@@ -1,1 +1,0 @@
-export { CSAT2Survey, type CSAT2SurveyProps } from './CSAT2Survey';