+++ src/surveys/NPS10Survey/index.ts
@@ -1,1 +1,0 @@
-export { NPS10Survey, type NPS10SurveyProps } from './NPS10Survey';