+++ src/surveys/Survey/index.ts
@@ -1,0 +1,8 @@
+export {
+  Survey,
+  type SurveyProps,
+  type CsatSurveyProps5,
+  type CsatSurveyProps2,
+  type CesSurveyProps,
+  type NpsSurveyProps
+} from './Survey';